+++ src/PhpWord/Shared/XMLWriter.php
@@ -58,11 +58,6 @@ class XMLWriter
      */
     public function __construct($tempLocation = self::STORAGE_MEMORY, $tempFolder = './')
     {
-        // Define date format
-        if (!defined('DATE_W3C')) {
-            define('DATE_W3C', 'Y-m-d\TH:i:sP');
-        }
-
         // Create internal XMLWriter
         $this->xmlWriter = new \XMLWriter();