+++ CHANGELOG.md
@@ -26,6 +26,7 @@ This release marked the change of PHPWord license from LGPL 2.1 to LGPL 3; new r
 - PDF Writer: Add TCPDF and mPDF as optional PDF renderer library - @ivanlanin
 - ODT Writer: Enable title element and custom document properties - @ivanlanin
 - ODT Reader: Ability to read standard and custom document properties - @ivanlanin
+- Word2007 Writer: Enable the missing custom document properties writer - @ivanlanin
 
 ### Bugfixes